Handle malformed post IDs
Bug: A bogus post from a booru server was returned from an API call with
a 13-figure ID. The post does not exist on the server. It is not
retrievable from the website. It's a bug in the API.

Obviously a better option would be to make IDs of type String or at
least Long, but that's a more invasive change that could perhaps be left
TODO.